Ajout du message voulez vous supprimer la photo

+++ client/src/components/CardChasseurs.tsx
@@ -28,6 +28,11 @@ function CardChasseurs() {
   }, []);
 
   const handleDelete = (photoId: number) => {
+    const isConfirmed = window.confirm(
+      "Êtes-vous sûr·e de vouloir supprimer cette photo ?",
+    );
+
+    if (!isConfirmed) return; // Si l'utilisateur annule, on stoppe la suppression
     fetch(`${import.meta.env.VITE_API_URL}/api/photos/${photoId}`, {
       method: "DELETE",
     })
